creates OrganizationType.swift
