+++ FullCircle/FullCircle/Models/Organization/OrganizationType.swift
@@ -1,0 +1,18 @@
+//
+//  OrganizationType.swift
+//  FullCircle
+//
+//  Created by Sunni Tang on 7/21/20.
+//  Copyright © 2020 FullCircle. All rights reserved.
+//
+
+import Foundation
+
+enum OrganizationType: String, Codable {
+    case movement
+    case nonprofit
+    case forProfit
+    case lobby
+    case academic
+    case individual
+}
